Make some stylistic changes to the "getstarted"
I have made some changes to the "getting started" tutorial, to improve the clarity and the style. An especially important point for me is that the definition of "mutation rate" and "crossover rate" is made clear (this has been an issue for me), and I have now attempted to add such definitions.